refactor(suite): extract sign verify package
