+++ packages/suite/package.json
@@ -123,6 +123,7 @@
         "@suite/router": "workspace:*",
         "@suite/sentry": "workspace:*",
         "@suite/settings": "workspace:*",
+        "@suite/sign-verify": "workspace:*",
         "@suite/suite-sync": "workspace:*",
         "@suite/thp": "workspace:*",
         "@suite/tor": "workspace:*",
