+++ packages/suite/src/actions/wallet/constants/index.ts
@@ -2,7 +2,6 @@ import { COINJOIN } from '@suite/coinjoin';
 
 import * as GRAPH from './graphConstants';
 import * as RECEIVE from './receiveConstants';
-import * as SIGN_VERIFY from './signVerifyConstants';
 import * as WALLET from './walletConstants';
 
-export { SIGN_VERIFY, RECEIVE, WALLET, GRAPH, COINJOIN };
+export { RECEIVE, WALLET, GRAPH, COINJOIN };
