+++ suite/sign-verify/jest.config.js
@@ -1,0 +1,6 @@
+const baseConfig = require('../../jest.config.base.swc');
+
+module.exports = {
+    ...baseConfig,
+    roots: ['<rootDir>/src', '<rootDir>/../../suite-common/test-utils/__mocks__'],
+};
