+++ suite/sign-verify/package.json
@@ -1,0 +1,47 @@
+{
+    "name": "@suite/sign-verify",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "test:unit": "yarn g:jest",
+        "type-check": "yarn g:tsc --build"
+    },
+    "dependencies": {
+        "@hookform/resolvers": "^5.4.0",
+        "@suite-common/address": "workspace:*",
+        "@suite-common/dependency-injection": "workspace:*",
+        "@suite-common/device": "workspace:*",
+        "@suite-common/receive": "workspace:*",
+        "@suite-common/suite-types": "workspace:*",
+        "@suite-common/toast-notifications": "workspace:*",
+        "@suite-common/validators": "workspace:*",
+        "@suite-common/wallet-config": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
+        "@suite-common/wallet-types": "workspace:*",
+        "@suite-common/wallet-utils": "workspace:*",
+        "@suite/account": "workspace:*",
+        "@suite/address": "workspace:*",
+        "@suite/device": "workspace:*",
+        "@suite/intl": "workspace:*",
+        "@trezor/components": "workspace:*",
+        "@trezor/connect": "workspace:*",
+        "@trezor/connect-common": "workspace:*",
+        "@trezor/dom-utils": "workspace:*",
+        "@trezor/icons": "workspace:*",
+        "@trezor/type-utils": "workspace:*",
+        "react": "19.2.3",
+        "react-hook-form": "^7.77.0",
+        "react-redux": "9.3.0",
+        "redux": "^5.0.1",
+        "redux-thunk": "^3.1.0"
+    },
+    "devDependencies": {
+        "@suite-common/test-utils": "workspace:*",
+        "@types/react": "19.2.14"
+    }
+}
