+++ suite/sign-verify/src/index.ts
@@ -1,0 +1,2 @@
+export { SignVerify } from './SignVerify';
+export { type SignVerifyAction } from './signVerifyActions';
