+++ suite/sign-verify/src/useCopySignedMessage.ts
@@ -1,8 +1,8 @@
+import { useDispatch } from 'react-redux';
+
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { type Network } from '@suite-common/wallet-config';
 import { copyToClipboard } from '@trezor/dom-utils';
-
-import { useDispatch } from 'src/hooks/suite';
 
 type SignedMessageData = {
     message?: string;
